+++ src/main/java/com/debateseason_backend_v1/domain/user/service/response/RegisterResponse.java
@@ -1,4 +1,4 @@
-package com.debateseason_backend_v1.domain.auth.service.response;
+package com.debateseason_backend_v1.domain.user.service.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
